+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/Config.java
@@ -48,6 +48,8 @@ public interface Config {
 
     SimpleEnergyUsageEntry getNetworkReceiver();
 
+    SimpleEnergyUsageEntry getNetworkTransmitter();
+
     interface SimpleEnergyUsageEntry {
         long getEnergyUsage();
     }