+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/ContentIds.java
@@ -51,6 +51,7 @@ public final class ContentIds {
     public static final ResourceLocation CONFIGURATION_CARD = createIdentifier("configuration_card");
     public static final ResourceLocation NETWORK_RECEIVER = createIdentifier("network_receiver");
     public static final ResourceLocation NETWORK_CARD = createIdentifier("network_card");
+    public static final ResourceLocation NETWORK_TRANSMITTER = createIdentifier("network_transmitter");
 
     private ContentIds() {
     }