+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/ContentNames.java
@@ -23,6 +23,7 @@ public final class ContentNames {
     public static final MutableComponent INTERFACE = createTranslation("block", "interface");
     public static final MutableComponent DISK_DRIVE = createTranslation("block", "disk_drive");
     public static final MutableComponent NETWORK_RECEIVER = createTranslation("block", "network_receiver");
+    public static final MutableComponent NETWORK_TRANSMITTER = createTranslation("block", "network_transmitter");
 
     private ContentNames() {
     }