+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/Tags.java
@@ -22,6 +22,7 @@ public final class Tags {
     public static final TagKey<Item> DESTRUCTORS = createTag("destructors");
     public static final TagKey<Item> WIRELESS_TRANSMITTERS = createTag("wireless_transmitters");
     public static final TagKey<Item> NETWORK_RECEIVERS = createTag("network_receivers");
+    public static final TagKey<Item> NETWORK_TRANSMITTERS = createTag("network_transmitters");
 
     private Tags() {
     }