+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/storage/storageblock/AbstractStorageBlockBlockEntity.java
@@ -176,4 +176,5 @@ abstract class AbstractStorageBlockBlockEntity<T>
         buf.writeLong(getNode().getCapacity());
         filter.getFilterContainer().writeToUpdatePacket(buf);
     }
+
 }