+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/datagen/BlockModelProvider.java
@@ -12,6 +12,11 @@ import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUti
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createIdentifier;
 
 public class BlockModelProvider extends net.minecraftforge.client.model.generators.BlockModelProvider {
+    private static final ResourceLocation EMISSIVE_ALL_CUTOUT = createIdentifier("block/emissive_all_cutout");
+    private static final ResourceLocation EMISSIVE_NORTH_CUTOUT = createIdentifier("block/emissive_north_cutout");
+    private static final ResourceLocation NORTH_CUTOUT = createIdentifier("block/north_cutout");
+    private static final ResourceLocation ALL_CUTOUT = createIdentifier("block/all_cutout");
+
     public BlockModelProvider(final PackOutput output, final ExistingFileHelper existingFileHelper) {
         super(output, MOD_ID, existingFileHelper);
     }
@@ -25,6 +30,7 @@ public class BlockModelProvider extends net.minecraftforge.client.model.generato
         registerDetectors();
         registerWirelessTransmitters();
         registerNetworkReceivers();
+        registerNetworkTransmitters();
     }
 
     private void registerCables() {
@@ -42,12 +48,11 @@ public class BlockModelProvider extends net.minecraftforge.client.model.generato
     }
 
     private void registerControllers() {
-        final ResourceLocation base = createIdentifier("block/emissive_all_cutout");
         final ResourceLocation off = createIdentifier("block/controller/off");
         final ResourceLocation on = createIdentifier("block/controller/on");
         Blocks.INSTANCE.getController().forEach((color, id, controller) -> {
             final ResourceLocation cutout = createIdentifier("block/controller/cutouts/" + color.getName());
-            withExistingParent("block/controller/" + color.getName(), base)
+            withExistingParent("block/controller/" + color.getName(), EMISSIVE_ALL_CUTOUT)
                 .texture("particle", off)
                 .texture("all", on)
                 .texture("cutout", cutout);
@@ -60,7 +65,7 @@ public class BlockModelProvider extends net.minecraftforge.client.model.generato
             registerEmissiveGrids(name, color.getName(), cutout);
         });
         final ResourceLocation inactiveCutout = createIdentifier("block/" + name + "/cutouts/inactive");
-        registerGrids(name, "inactive", inactiveCutout, createIdentifier("block/north_cutout"));
+        registerGrids(name, "inactive", inactiveCutout, NORTH_CUTOUT);
     }
 
     private void registerGrids(final String name,
@@ -85,7 +90,7 @@ public class BlockModelProvider extends net.minecraftforge.client.model.generato
     }
 
     private void registerEmissiveGrids(final String name, final String variantName, final ResourceLocation cutout) {
-        registerGrids(name, variantName, cutout, createIdentifier("block/emissive_north_cutout"));
+        registerGrids(name, variantName, cutout, EMISSIVE_NORTH_CUTOUT);
     }
 
     private void registerDetectors() {
@@ -113,18 +118,36 @@ public class BlockModelProvider extends net.minecraftforge.client.model.generato
     }
 
     private void registerNetworkReceivers() {
-        final ResourceLocation emissiveBase = createIdentifier("block/emissive_all_cutout");
         final ResourceLocation baseTexture = createIdentifier("block/network_receiver/base");
         Blocks.INSTANCE.getNetworkReceiver().forEach((color, id, receiver) -> {
             final ResourceLocation cutout = createIdentifier("block/network_receiver/cutouts/" + color.getName());
-            withExistingParent("block/network_receiver/" + color.getName(), emissiveBase)
+            withExistingParent("block/network_receiver/" + color.getName(), EMISSIVE_ALL_CUTOUT)
                 .texture("particle", baseTexture)
                 .texture("all", baseTexture)
                 .texture("cutout", cutout);
         });
-        withExistingParent("block/network_receiver/inactive", createIdentifier("block/all_cutout"))
+        withExistingParent("block/network_receiver/inactive", ALL_CUTOUT)
             .texture("particle", baseTexture)
             .texture("all", baseTexture)
             .texture("cutout", createIdentifier("block/network_receiver/cutouts/inactive"));
+    }
+
+    private void registerNetworkTransmitters() {
+        final ResourceLocation baseTexture = createIdentifier("block/network_transmitter/base");
+        Blocks.INSTANCE.getNetworkTransmitter().forEach((color, id, receiver) -> {
+            final ResourceLocation cutout = createIdentifier("block/network_transmitter/cutouts/" + color.getName());
+            withExistingParent("block/network_transmitter/" + color.getName(), EMISSIVE_ALL_CUTOUT)
+                .texture("particle", baseTexture)
+                .texture("all", baseTexture)
+                .texture("cutout", cutout);
+        });
+        withExistingParent("block/network_transmitter/inactive", ALL_CUTOUT)
+            .texture("particle", baseTexture)
+            .texture("all", baseTexture)
+            .texture("cutout", createIdentifier("block/network_transmitter/cutouts/inactive"));
+        withExistingParent("block/network_transmitter/error", EMISSIVE_ALL_CUTOUT)
+            .texture("particle", baseTexture)
+            .texture("all", baseTexture)
+            .texture("cutout", createIdentifier("block/network_transmitter/cutouts/error"));
     }
 }