+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/datagen/ItemModelProvider.java
@@ -13,6 +13,7 @@ import com.refinedmods.refinedstorage2.platform.common.grid.GridBlock;
 import com.refinedmods.refinedstorage2.platform.common.importer.ImporterBlock;
 import com.refinedmods.refinedstorage2.platform.common.networking.CableBlock;
 import com.refinedmods.refinedstorage2.platform.common.networking.NetworkReceiverBlock;
+import com.refinedmods.refinedstorage2.platform.common.networking.NetworkTransmitterBlock;
 import com.refinedmods.refinedstorage2.platform.common.storage.externalstorage.ExternalStorageBlock;
 import com.refinedmods.refinedstorage2.platform.common.wirelesstransmitter.WirelessTransmitterBlock;
 
@@ -25,6 +26,9 @@ import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUti
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createIdentifier;
 
 public class ItemModelProvider extends net.minecraftforge.client.model.generators.ItemModelProvider {
+    private static final String CUTOUT_TEXTURE_KEY = "cutout";
+    private static final String CABLE_TEXTURE_KEY = "cable";
+
     public ItemModelProvider(final PackOutput output, final ExistingFileHelper existingFileHelper) {
         super(output, MOD_ID, existingFileHelper);
     }
@@ -44,6 +48,7 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         registerDestructors();
         registerWirelessTransmitters();
         registerNetworkReceivers();
+        registerNetworkTransmitters();
     }
 
     private void registerCables() {
@@ -52,7 +57,7 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         blocks.forEach((color, id, block) -> singleTexture(
             id.getPath(),
             base,
-            "cable",
+            CABLE_TEXTURE_KEY,
             createIdentifier("block/cable/" + color.getName()))
         );
     }
@@ -63,7 +68,7 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         blocks.forEach((color, id, block) -> singleTexture(
             id.getPath(),
             base,
-            "cable",
+            CABLE_TEXTURE_KEY,
             createIdentifier("block/cable/" + color.getName())
         ));
     }
@@ -74,7 +79,7 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         blocks.forEach((color, id, block) -> singleTexture(
             id.getPath(),
             base,
-            "cable",
+            CABLE_TEXTURE_KEY,
             createIdentifier("block/cable/" + color.getName())
         ));
     }
@@ -85,7 +90,7 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         blocks.forEach((color, id, block) -> singleTexture(
             id.getPath(),
             base,
-            "cable",
+            CABLE_TEXTURE_KEY,
             createIdentifier("block/cable/" + color.getName())
         ));
     }
@@ -156,7 +161,7 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         blocks.forEach((color, id, block) -> singleTexture(
             id.getPath(),
             base,
-            "cable",
+            CABLE_TEXTURE_KEY,
             createIdentifier("block/cable/" + color.getName())
         ));
     }
@@ -167,7 +172,7 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         blocks.forEach((color, id, block) -> singleTexture(
             id.getPath(),
             base,
-            "cable",
+            CABLE_TEXTURE_KEY,
             createIdentifier("block/cable/" + color.getName())
         ));
     }
@@ -178,7 +183,7 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         blocks.forEach((color, id, block) -> singleTexture(
             id.getPath(),
             base,
-            "cutout",
+            CUTOUT_TEXTURE_KEY,
             createIdentifier("block/wireless_transmitter/cutouts/" + color.getName())
         ));
     }
@@ -189,8 +194,19 @@ public class ItemModelProvider extends net.minecraftforge.client.model.generator
         blocks.forEach((color, id, block) -> singleTexture(
             id.getPath(),
             base,
-            "cutout",
+            CUTOUT_TEXTURE_KEY,
             createIdentifier("block/network_receiver/cutouts/" + color.getName())
+        ));
+    }
+
+    private void registerNetworkTransmitters() {
+        final ResourceLocation base = createIdentifier("block/network_transmitter/inactive");
+        final ColorMap<NetworkTransmitterBlock> blocks = Blocks.INSTANCE.getNetworkTransmitter();
+        blocks.forEach((color, id, block) -> singleTexture(
+            id.getPath(),
+            base,
+            CUTOUT_TEXTURE_KEY,
+            createIdentifier("block/network_transmitter/cutouts/" + color.getName())
         ));
     }